Initial Commit
Initial commit to this repository. I am going through the implementation that is hosted here https://github.com/pierluigiferrari/ssd_keras, and I am trying to build a working minimum with some minor changes of my own in order to get a low-level understanding of how the SSD algorithm works.